Allow Python 3.6 to fail
I'm not happy with this.

There's some kind of version compatibility issue - 2.7 parses things
fine, 3.6 does not.

However for now, 2.7 is the canonical version (because the Pi runs 2.7
and it's only a Pi on which I'm willing to do all the BOOTP config for
Monitor communication) and so, while I care about 3.6, it's not a
blocking concern