Rewrote association code to use Packets, plus convenience function
From the PIPG:

"""
A Computer Client can use the pre-defined building blocks for the
Session Data, Presentation Header, and Presentation Trailer listed in
the appendix to conveniently build valid messages (“Association Control
Protocol Examples” on page 298 for a list of building blocks). Only the
User Data block of the Association Request must be filled with Computer
Client-specific data.
"""

Unfortunately this also means I can't find structures for some of the
segments, hence the hardcoded defaults here...